+++ lib/weather_app.dart
@@ -1,0 +1,65 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_training/components/action_button.dart';
+import 'package:flutter_training/components/weather_information.dart';
+
+class WeatherApp extends StatelessWidget {
+  const WeatherApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(
+          seedColor: Colors.blue,
+        ),
+        brightness: Brightness.light,
+        useMaterial3: true,
+      ),
+      darkTheme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(
+          seedColor: Colors.blue,
+          brightness: Brightness.dark,
+        ),
+        useMaterial3: true,
+      ),
+      home: Scaffold(
+        // 真ん中で2:1の領域を作る
+        body: Center(
+          child: FractionallySizedBox(
+            widthFactor: 0.5,
+            child: Column(
+              children: [
+                const Spacer(),
+                const WeatherInformation(),
+                Expanded(
+                  child: Column(
+                    // SizedBoxを用いて
+                    // WeatherInformationとActionButtonの隙間を作る
+                    children: [
+                      const SizedBox(
+                        height: 80,
+                      ),
+                      Row(
+                        mainAxisAlignment: MainAxisAlignment.spaceAround,
+                        children: [
+                          ActionButton(
+                            buttonTitle: 'close',
+                            onPressed: () {},
+                          ),
+                          ActionButton(
+                            buttonTitle: 'reload',
+                            onPressed: () {},
+                          ),
+                        ],
+                      ),
+                    ],
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
